gas: loongarch: allow to assemble UTF-8 encoded Unicode names
Move is_*label* from libopcode to gas, and adjust it to use macros
defined in read.h to check if a label is legal.

libopcode is designed for disassemble, put a lot of code for assembler
(not disassembler) there is bad.

This should fix gcc.dg/ucnid-*.c failures in GCC.